Refactor how the worklet receives audio data. This fixes 99% of any leaky heap memory.
Instead of re-creating the assembly module every time we start a new song, we instead create it once, onMount for the component. When a user clicks on a song to play, that audio data is instead passed to the already existing worklet via ports. (The worklet then cleans up the previous module/buffer.)